% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.results.R
\name{CEN-Methods}
\alias{CEN-Methods}
\alias{summary.CEN}
\alias{print.CEN}
\alias{plot.CEN}
\title{Summary and Plot Results}
\usage{
\method{summary}{CEN}(object, ...)

\method{print}{CEN}(x, ...)

\method{plot}{CEN}(x, what = c("strength", "network", "edge", "centrality"), ...)
}
\arguments{
\item{object}{output of '\link{CompareEdgelistNetworks}', object of class "CEN"}

\item{...}{Arguments only used in plot method}

\item{x}{output of '\link{CompareEdgelistNetworks}', object of class "CEN"}

\item{what}{defines what has to be plotted: results pertaining to test on invariance of global strength ("strength"), network structure ("network"), edge strength ("edge"), or specific centrality measure ("centrality")}
}
\value{
CEN-Methods contains three functions to display the output from '\link{CompareEdgelistNetworks}'.
\code{\link{summary.CEN}} summarizes the test statistics of the compared networks in detail.
\code{\link{print.CEN}} briefly summarizes the test statistics of the compared networks.
\code{\link{plot.CEN}} graphically presents the significance tests of the various variance measures in form of histograms.
}
\description{
This function can summary, print and plot results from '\link{CompareEdgelistNetworks}'
}
