% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareEdgelistNetworks.R
\name{NetworkComparr}
\alias{NetworkComparr}
\title{Statistical Comparison of Networks based on Edgelists}
\description{
A permutation-based hypothesis test for statistical comparison of two networks based on the invariance measures of the R package 'NetworkComparisonTest' by van Borkulo et al. (2022): global strength invariance, network structure invariance, edge invariance, and various centrality measures. Edgelists of two comparable group networks from dependent or independent samples are used as input. These edgelists can be generated from individual graphical network representations, such as concept maps. The networks can be directed or undirected.

Keywords: concept-maps, edgelists, comparing networks, permutation-test
}
\references{
Salmaso, L., & Pesarin, F. (2010). Permutation tests for complex data: theory, applications and software. John Wiley & Sons. https://doi.org/10.1002/9780470689516

van Borkulo, C. D., van Bork, R., Boschloo, L., Kossakowski, J. J., Tio, P., Schoevers, R. A., Borsboom, D., & Waldorp, L. J. (2022). Comparing network structures on three aspects: A permutation test. Psychological methods. https://doi.org/10.1037/met0000476.
}
\author{
Lara Trani \href{mailto:lara.trani@rptu.de}{lara.trani@rptu.de}
Maike Sauer \href{mailto:maike.sauer@rptu.de}{maike.sauer@rptu.de}
}
\keyword{comparing-networks}
\keyword{concept-maps}
\keyword{edgelists}
\keyword{permutation-test}
